% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade.R
\name{DimExpansion}
\alias{DimExpansion}
\title{Main fitting procedure.}
\usage{
DimExpansion(X, Y, lambda, Maxd, EmpSemivariogram, para.initial,
  Method = "OLS", model = "Exponential")
}
\arguments{
\item{X}{The records of location coordinates, a \eqn{n \times d} matrix,
where \eqn{n} is the number of locations, and \eqn{d} is the dimension of the record location.
Usually, \eqn{d =2} (latitude and longitude coordinates).}

\item{Y}{Realizations of the Gaussian process \eqn{Y(X)}.}

\item{lambda}{A prespecified tuning parameter, \eqn{\lambda >0}.}

\item{Maxd}{A prespecified maximum coordinate dimension of each location.}

\item{EmpSemivariogram}{The empirical semivariogram that is computed by \code{\link{EmpSemivariogramFit}}.}

\item{para.initial}{Initial parameters of \eqn{(\phi,Z)} in the objective functions, \eqn{f_{OLS},f_{WLS},f_{GLS}}.}

\item{Method}{Three optional methods, Method = c("OLS","WLS","GLS"), which corresponds to
the Least-squares, weighted least-squares, and generalized least-squares fitting procedures (default:OLS).}

\item{model}{Semivariogram model, it has three options, "Exponential", "Power","Gaussian" (default: "Exponential").}
}
\value{
Return the estimates of \eqn{\phi} and locations coordinates \eqn{Z} of the learned dimensions.
}
\description{
The fitting procedure is to minimize one of the objective functions,
\eqn{f_{OLS},f_{WLS},f_{GLS}}.
Its purpose is to estimate the semivariogram function and the coordinates of the learned dimensions.
The method applied in each iteration is BFGS (Broyden,1979).
}
\examples{
n = 10  # number of locations.
Maxd = 3
SimRep = 1000 # number of observations for each location
lambda = 0.01
trueTheta = 10
set.seed(123)
Data = GenerateData(n=n, SimRep = SimRep,trueTheta = trueTheta)
Y = Data$YObs
X = Data$LocIndex[,1:2]
EmpSemivariogram = EmpSemivariogramFit(Y)
para.initial = c(1,1,Data$LocIndex[,-c(1:2)])
DimExpansion(X,Y,lambda,Maxd,EmpSemivariogram, para.initial, Method="OLS",model = "Exponential")
DimExpansion(X,Y,lambda,Maxd,EmpSemivariogram, para.initial, Method="GLS",model = "Exponential")
DimExpansion(X,Y,lambda,Maxd,EmpSemivariogram, para.initial, Method="WLS",model = "Exponential")
}
\references{
\itemize{
 \item Bornn, L.,Shaddick, G., and Zidek, J.V. (2012). Modeling non-stationary processes through dimension expansion.
   Journal of American Statistical Association, 107:281–289.
 \item Broyden, C.G.(1979). The convergence of a class of double-rank minimization algorithms.
   Journal of the Institute of Mathematics and Its Application,6:76-90.
 \item Shanshan Qin, Bin Sun, Yuehua Wu, Yuejiao Fu. (2020)
Generalized Least-Squares in Dimension Expansion Method for Nonstationary Processes.
}
}
