% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade.R
\name{EmpSemivariogramFit}
\alias{EmpSemivariogramFit}
\title{Estimate the empirical semivariogram function.}
\usage{
EmpSemivariogramFit(Y)
}
\arguments{
\item{Y}{Realizations of the spatial random process \eqn{Y(X)};}
}
\value{
A vector formed by concatenating the columns of upper triangular of the empirical semivariogram.
}
\description{
This is to estimate the empirical semivariogram function.
}
\references{
Shanshan Qin, Bin Sun, Yuehua Wu, Yuejiao Fu.(2020)
Generalized Least-Squares in Dimension Expansion Method for Nonstationary Processes.
}
