% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade.R
\name{Gamma_phiFun}
\alias{Gamma_phiFun}
\title{The Isotropic Semivariogram Model.}
\usage{
Gamma_phiFun(phi, x, model = "Exponential")
}
\arguments{
\item{phi}{Two dimensional vector-valued parameters, \eqn{\phi_1, \phi_2 \geq 0}.}

\item{x}{The Euclidean distance of two locations, \eqn{x \geq 0}.}

\item{model}{The semivariogram model, it has three options,
"Exponential", "Power", "Gaussian" (default:"Exponential").}
}
\value{
A scalar representing the spatial isotropic variogram estimate for x.
}
\description{
Define a semivariogram function (no nugget effect).
Herein, we provide three options, "Exponential", "Power","Gaussian".
}
\references{
Shanshan Qin, Bin Sun, Yuehua Wu, Yuejiao Fu.(2020)
Generalized Least-Squares in Dimension Expansion Method for Nonstationary Processes.
}
