% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spade.R
\name{ObjectiveFunBorrn}
\alias{ObjectiveFunBorrn}
\title{The objective function in Bornn's paper.}
\usage{
ObjectiveFunBorrn(X, para, lambda, Maxd, EmpSemivariogram, model)
}
\arguments{
\item{X}{The records of location coordinates, a \eqn{n \times d} matrix,
where \eqn{n} is the number of locations, and \eqn{d} is the dimension of the record location.
Usually, \eqn{d =2} (latitude and longitude coordinates).}

\item{para}{A vector-valued parameters \eqn{(\phi,Z)} in the objective function, dimension (Maxd-2)n + 2.}

\item{lambda}{A prespecified tuning parameter, \eqn{\lambda >0}.}

\item{Maxd}{A prespecified maximum coordinate dimension of each location.}

\item{EmpSemivariogram}{The empirical semivariogram that is computed by \code{\link{EmpSemivariogramFit}}.}

\item{model}{The semivariogram model, it has three options, "Exponential", "Power","Gaussian" (default:"Exponential").}
}
\value{
An objective value
}
\description{
The OLS fitting procedure is to minimize the objective function,
\deqn{f_{OLS}
= \sum\limits_{j<i} \{ \hat{\gamma}_{i,j}-\gamma_{\phi} d_{i,j}([X,Z])\}^{2}
+ \lambda\sum\limits_{k=1}^{p}\|Z_{.k}\|_{1} }
}
\references{
Bornn, L.,Shaddick, G., and Zidek, J.V.(2012). Modeling non-stationary processes through dimension expansion.
   Journal of American Statistical Association, 107:281–289.
}
